backend/requirements.txt
fastapi
uvicorn[standard]
python-multipart
pydantic
PyPDF2
pdfplumber
reportlab
openai
whisper
SQLAlchemy
pytesseract
Pillow
gTTS
transformers
torch
